debugged read cmmd test error with Scott. refactored commands into single file and their respective tests. working on debugging create command when cache is nil
